requirements.txt
# === 数据源 ===
akshare>=1.18.0

# === 数值计算 ===
pandas>=2.1.0
numpy>=1.24.0

# === 技术指标 ===
ta>=0.11.0

# === 预测模型 ===
torch>=2.1.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# === 情感分析 ===
transformers>=4.35.0
sentencepiece>=0.1.99

# === Web 后端 ===
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic-settings>=2.1.0

# === Web 前端 ===
streamlit>=1.28.0
plotly>=5.18.0

# === 数据库 ===
sqlalchemy>=2.0.23
aiosqlite>=0.19.0

# === 调度器 ===
apscheduler>=3.10.4

# === 工具 ===
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0

# === 测试 ===
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
